refactor: 컴포넌트 구조 변경
컴포넌트 재사용을 위해 구조 변경 및 부트스트랩 CSS 적용
